refact: remove Stop
use context
